requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.5
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.3.0
azure-ai-inference==1.0.0b9
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.35.1
billiard==4.2.2
blinker==1.9.0
cachetools==6.2.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.116.1
filelock==3.19.1
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.1
humanfriendly==10.0
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
langchain==1.0.7
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.5
langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
limits==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.1
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.5.0
networkx==3.5
numpy==2.3.3
onnx==1.19.0
onnxruntime
openai==2.7.2
openpyxl==3.1.5
optimum[onnxruntime]
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.3
psycopg==3.2.10
psycopg-binary==3.2.10
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
# pyreadline3==3.5.4
pytest==8.4.2
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.3
redis==5.3.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.4
toml==0.10.2
torch
tornado==6.5.2
tqdm==4.67.1
transformers
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.22.0
vaderSentiment==3.3.2
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.14
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
pgvector
